# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: IceCream
  description: ''
  version: 1.0.0
servers:
  -
    url: 'https://icecream'
paths:
  /api/caterogies/new:
    post:
      summary: 'create caterogy'
      operationId: createCaterogy
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Caterogies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: ''
                  example: est
              required:
                - name
      security: []
...
  /api/caterogies/edit:
    put:
      summary: 'update caterogy'
      operationId: updateCaterogy
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Caterogies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  description: ''
                  example: 3
                name:
                  type: string
                  description: nullable
                  example: tempora
              required:
                - id
      security: []
...
  /api/products/edit:
    put:
      summary: 'Update product'
      operationId: updateProduct
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Products
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  description: 'The id of the product.'
                  example: 1
                caterogy_id:
                  type: integer
                  description: 'nullable The id of the caterogy.'
                  example: 1
                name:
                  type: string
                  description: 'nullable The name of the product.'
                  example: 'Product 1'
                price:
                  type: integer
                  description: 'nullable The price of the product.'
                  example: 100
                count:
                  type: integer
                  description: 'nullable The count of the product.'
                  example: 100
                type_id:
                  type: integer
                  description: 'nullable The id of the type.'
                  example: 1
                is_active:
                  type: boolean
                  description: 'nullable The status of the product.'
                  example: true
              required:
                - id
      security: []
...